+++ pkg/internal/tools/doc-gen-helper/model.go
@@ -1,0 +1,17 @@
+package main
+
+type DeprecatedResourcesContext struct {
+	Resources []DeprecatedResource
+}
+
+type DeprecatedResource struct {
+	Name string
+}
+
+type DeprecatedDatasourcesContext struct {
+	Datasources []DeprecatedDatasource
+}
+
+type DeprecatedDatasource struct {
+	Name string
+}
